SRTN basics implemented not correct
There are surely some bugs in this implementation
The queue is not actually getting the shortest remaining time or maybe,
because of the multithreading, I'm stopping a task then continuing the
same task instead of starting a new one (maybe try getting a higher
duration difference will check it out better.

Also, a test with context change (useful for SRTN) was added.