Test data, charts and chart images
I've executed the testing required by the specification and collected
the data necessary and created charts, as requested.

For the creation of the charts, the javascript library Highcharts was
used. Afterwise, every chart was exported to jpg.

For all of this work, the helpful cpp code calcstuff.cpp was created.

The data inside the charts files was hardcoded.