Recreating Page and using pagination
- unsigned get_page(unsigned block) was erased
- table is an unordered_map, not a map
- many new functions created
- static functions are all in Page, not Page::Algorithm (I should do this to Memory too)
- interfaces for accessing memory
- identifying necessity for page substitution and calling page substitution algorithms (stuff from Memory::Algorithm) on the right places

So, basically, we just need to implement the functions page_to_remove and signal for the four desired algorithms to fill Page::Algorithm and we are one
Also, maybe Memory needs some serious refactoring, there's actually a function (Memory::visit) which became just a wrapper for Page::visit. Some functions on Page are sort of badly named, also.